get configurations from publisher yml
