+++ app/client/render.js
@@ -7,6 +7,7 @@ import { Footer } from "../isomorphic/components/layouts/footer";
 import { NavbarSearch } from "../isomorphic/components/layouts/header/navbar-search";
 import { NavBar } from "../isomorphic/components/layouts/header/nav-bar";
 import { TopAd } from "../isomorphic/components/ads/top-ad";
+import { PlaceholderGenerator } from "../isomorphic/components/placeholder-generator";
 
 export function preRenderApplication(store) {
   const hydrate = { hydrate: !global.qtLoadedFromShell };
@@ -28,8 +29,11 @@ export function preRenderApplication(store) {
 // This is a separate file as everything from here on is hot reloaded when the app changes
 export function renderApplication(store) {
   const enableAds = get(store.getState(), ["qt", "config", "ads-config", "dfp_ads", "enable_ads"]);
+  const enablePlaceholder = get(store.getState(), ["qt", "config", "publisher-attributes", "enable_placeholder"]);
+  global.showPlaceholder = enablePlaceholder;
 
   enableAds && renderComponent(TopAd, "top-ad", store);
+  enablePlaceholder && renderComponent(PlaceholderGenerator, "placeholder-generator", store);
   renderIsomorphicComponent("container", store, pickComponent, {
     hydrate: !global.qtLoadedFromShell
   });
