+++ app/server/handlers/render-layout.js
@@ -10,6 +10,7 @@ import fontFace from "../font";
 import { BreakingNewsView } from "../../isomorphic/components/breaking-news-view";
 import { TopAd } from "../../isomorphic/components/ads/top-ad";
 import { getConfig, extractor, getCriticalCss, getArrowCss } from "../helpers";
+import get from "lodash.get";
 
 const cssContent = assetPath("app.css") ? readAsset("app.css") : "";
 const fontJsContent = assetPath("font.js") ? readAsset("font.js") : "";
@@ -29,6 +30,13 @@ export async function renderLayout(res, params) {
   const chunk = params.shell ? null : allChunks[getChunkName(params.pageType)];
   const criticalCss = await getCriticalCss();
   const arrowCss = await getArrowCss(params.store.getState());
+
+  const enablePlaceholder = get(params.store.getState(), [
+    "qt",
+    "config",
+    "publisher-attributes",
+    "enable_placeholder"
+  ]);
 
   res.render(
     "pages/layout",
@@ -65,7 +73,8 @@ export async function renderLayout(res, params) {
         isOnesignalEnable,
         oneSignalScript: params.oneSignalScript,
         enableAds,
-        loadAdsSynchronously
+        loadAdsSynchronously,
+        enablePlaceholder
       },
       params
     )
