+++ views/pages/layout.ejs
@@ -10,6 +10,18 @@
     <link rel="preconnect" href="https://prod-analytics.qlitics.com" crossorigin />
     <link rel="preconnect" href="https://fonts.gstatic.com/" crossorigin />
 
+    <script type="text/javascript">
+      window.GUMLET_CONFIG = {
+        hosts: [{ current:  "<%= cdnImage %>" , gumlet: "<%= cdnImage %>"}],
+        lazy_load: true,
+        auto_webp: true
+      }
+    </script>
+
+    <%_ if(!enablePlaceholder) { _%>
+      <script src="https://cdn.gumlet.com/gumlet.js/2.0/gumlet.min.js" type="text/javascript" defer="defer"></script>
+    <%_ } _%>
+
     <%_ if(enableAds && loadAdsSynchronously) { _%>
       <script async="async" src="https://www.googletagservices.com/tag/js/gpt.js"></script>
     <%_ } _%>
@@ -53,6 +65,7 @@
   </head>
 
   <body>
+    <div id='placeholder-generator'></div>
     <% if (isGtmEnable) { %>
       <%- include ./partials/noscriptgtm -%>
     <% } %>
